Artifactory setup in Repository Component
