+++ VERSION
@@ -1,1 +1,1 @@
-1.10.0
+1.11.0
