+++ skelebot/components/__init__.py
@@ -1,3 +1,3 @@
 """Built-In Components of Skelebot and the Component Factory to Access Them"""
 
-from . import artifactory, bump, componentFactory, dexec, jupyter, kerberos, plugin, prime
+from . import artifactory, bump, componentFactory, dexec, jupyter, kerberos, plugin, prime, registry
