+++ skelebot/components/repository/__init__.py
@@ -1,0 +1,3 @@
+"""Scripts for the Repository Component"""
+
+from . import artifactoryRepo, artifactRepo, repository, s3fsRepo
