+++ skelebot/components/repository/repository.py
@@ -1,0 +1,140 @@
+"""Repository Component"""
+
+from schema import Schema, And, Optional, SchemaError
+from .artifactRepo import ArtifactRepo
+from .s3fsRepo import S3fsRepo
+from .artifactoryRepo import ArtifactoryRepo
+from ...objects.component import Activation, Component
+from ...objects.skeleYaml import SkeleYaml
+
+ERROR_NOT_COMPATIBLE = "No Compatible Version Found"
+ERROR_ALREADY_PUSHED = "This artifact version already exists. Please bump the version or use the force parameter (-f) to overwrite the artifact."
+
+class Artifact(SkeleYaml):
+    """
+    Artifact Class
+
+    Object contained within a list of the artifactory configuration in order to specify the
+    artifact files and artifact names
+    """
+
+    VERSIONED_NAME = "{filename}_v{version}.{ext}"
+
+    schema = Schema({
+        'name': And(str, error='Artifact \'name\' must be a String'),
+        'file': And(str, error='Artifact \'file\' must be a String'),
+    }, ignore_extra_keys=True)
+
+    name = None
+    file = None
+
+    def __init__(self, name, file):
+        """ Initialize the Artifact object with a name to be used in commands a file path """
+        self.name = name
+        self.file = file
+
+    def getVersionedName(self, version):
+        """ Constuct the full versioned filename for the artifact with the given version number """
+        ext = self.file.split(".")[1]
+        return self.VERSIONED_NAME.format(filename=self.name, version=version, ext=ext)
+
+class Repository(Component):
+    """
+    Repository Component Class
+
+    Provides the ability to push and pull artifacts that are defined in the skelebot config
+    file to and from Artifactory or S3 based on the project version number
+    """
+
+    activation = Activation.CONFIG
+    commands = ["push", "pull"]
+
+    schema = Schema({
+        Optional('artifacts'): And(list, error='Repository \'artifacts\' must be a List'),
+        Optional('s3fs'): And(dict, error='Repository \'s3fs\' must be an Object'),
+        Optional('artifactory'): And(dict, error='Repository \'artifactory\' must be an Object'),
+    }, ignore_extra_keys=True)
+
+    artifacts = None
+    s3fs = None
+    artifactory = None
+
+    @classmethod
+    def load(cls, config):
+        """Instantiate the Repository Class Object based on a config dict"""
+
+        cls.validate(config)
+
+        s3fs = None
+        artifactory = None
+        if ("s3fs" in config):
+            s3fs = S3fsRepo.load(config["s3fs"])
+        elif ("artifactory" in config):
+            artifactory = ArtifactoryRepo.load(config["artifactory"])
+        else:
+            raise SchemaError(None, "Repository must contain 's3fs' or 'artifactory' config")
+
+        artifactDicts = config["artifacts"]
+        artifacts = []
+        for artifact in artifactDicts:
+            newArtifact = Artifact.load(artifact)
+            artifacts.append(newArtifact)
+
+        return cls(artifacts, s3fs, artifactory)
+
+    def __init__(self, artifacts=None, s3fs=None, artifactory=None):
+        """Instantiate the Repository Class Object based on the provided parameters"""
+
+        self.artifacts = artifacts if artifacts is not None else []
+        self.s3fs = s3fs
+        self.artifactory = artifactory
+
+    def requiresPassword(self):
+        return self.artifactory is not None
+
+    def addParsers(self, subparsers):
+        """
+        SkeleParser Hook
+
+        Adds the parsers for push and pull commands to the SkeleParser to push artifacts
+        with the project version and pull artifacts with the provided version number
+        """
+
+        artifactNames = []
+        for artifact in self.artifacts:
+            artifactNames.append(artifact.name)
+
+        parser = subparsers.add_parser("push", help="Push an artifact to artifactory")
+        parser.add_argument("artifact", choices=artifactNames)
+        parser.add_argument("-f", "--force", action='store_true', help="Force the push")
+        if (self.requiresPassword()):
+            parser.add_argument("-u", "--user", help="Auth user for Artifactory")
+            parser.add_argument("-t", "--token", help="Auth token for Artifactory")
+
+        parser = subparsers.add_parser("pull", help="Pull an artifact from artifactory")
+        parser.add_argument("artifact", choices=artifactNames)
+        parser.add_argument("version", help="The version of the artifact to pull")
+        parser.add_argument("-o", "--override", action='store_true',  help="Override the model in the existing directory")
+        if (self.requiresPassword()):
+            parser.add_argument("-u", "--user", help="Auth user for Artifactory")
+            parser.add_argument("-t", "--token", help="Auth token for Artifactory")
+
+        return subparsers
+
+    def execute(self, config, args):
+
+        # Obtain the artifact that matches the provided name
+        selectedArtifact = None
+        for artifact in self.artifacts:
+            if (artifact.name == args.artifact):
+                selectedArtifact = artifact
+
+        ext = selectedArtifact.file.split(".")[1]
+
+        # Obtain the configured artifact repository
+        artifactRepo = self.s3fs if self.s3fs is not None else self.artifactory
+
+        if (args.job == "push"): # Push from Disk to Repo
+            artifactRepo.push(selectedArtifact, config.version, args.force, args.user, args.token)
+        elif (args.job == "pull"): # Pull from Repo to Disk
+            artifactRepo.pull(selectedArtifact, args.version, config.version, args.override, args.user, args.token)
