+++ skelebot/objects/semver.py
@@ -11,15 +11,18 @@ class Semver():
     patch = None
 
     def __init__(self, version):
+        """ Initialize the Major, Minor, and Patch numbers of a semantic version number """
         parts = version.split(".")
         self.major = parts[0]
         self.minor = parts[1]
         self.patch = parts[2]
 
     def __str__(self):
+        """ When converting to a String the format should read as major.minor.patch """
         return "{maj}.{min}.{pat}".format(maj=self.major, min=self.minor, pat=self.patch)
 
     def __lt__(self, other):
+        """ Check if this semver is smaller (less recent) than another semver """
         lt = False
         if (self.major < other.major) or ((self.major == other.major) and ((self.minor < other.minor) or ((self.minor == other.minor) and (self.patch < other.patch)))):
             lt = True
@@ -27,6 +30,7 @@ class Semver():
         return lt
 
     def __eq__(self, other):
+        """ Check if this semver is exactly equal to another semver """
         eq = True
         if (self.patch != other.patch) or (self.minor != other.minor) or (self.major != other.major):
             eq = False
@@ -34,6 +38,7 @@ class Semver():
         return eq
 
     def __le__(self, other):
+        """ Check if this semver is larger (more recent) than another semver """
         le = False
         if (self < other) or (self == other):
             le = True
@@ -41,7 +46,12 @@ class Semver():
         return le
 
     def isBackwardCompatible(self, semver):
-        """Determine whether this Semver backword compatible with the Semver provided"""
+        """
+        Determine whether this Semver is backward compatible with the Semver provided
+
+        For this semver to be backward compatible with another the Major version of each must be
+        the same and this semver must be at least as large (at least as recent) as the one provided
+        """
 
         compat = False
         if (self.major == semver.major) and (self >= semver):
